+++ src/back-end/database/db_build.sql
@@ -2,7 +2,6 @@ BEGIN;
 
 DROP TABLE IF EXISTS program CASCADE;
 DROP TABLE IF EXISTS content_package CASCADE;
-DROP TABLE IF EXISTS group CASCADE;
 DROP TABLE IF EXISTS department CASCADE;
 DROP TABLE IF EXISTS work_team CASCADE;
 DROP TABLE IF EXISTS employee CASCADE;
