+++ src/front-end/MyContentPackages.css
@@ -1,20 +1,20 @@
 body {
- background-color: #e6e6ff;
+  background-color: #e6e6ff;
 }
 .boxButton {
- margin-left: 58%;
- margin-top: 29%;
- text-align: center;
- width: 60px;
- height: 60px;
- background-color: #19334d;
- color: white;
- font: Arial;
- font-weight: bold;
- font-size: 15px;
- border-style: none;
- border-radius: 50%;
+  margin-left: 58%;
+  margin-top: 29%;
+  text-align: center;
+  width: 60px;
+  height: 60px;
+  background-color: #19334d;
+  color: white;
+  font: Arial;
+  font-weight: bold;
+  font-size: 15px;
+  border-style: none;
+  border-radius: 50%;
 }
 .boxImage {
- width: 38px;
+  width: 38px;
 }
