+++ src/front-end/buttons/allContentButtons.css
@@ -1,6 +1,6 @@
 .add {
-  margin-left: 15%;
-  margin-top: 2%;
+  margin-left: 18%;
+  margin-top: 16%;
   text-align: center;
   width: 120px;
   height: 30px;
@@ -59,21 +59,16 @@
   font-size: 15px;
   border-style: none;
 }
-.boxButton {
-  margin-left: 90%;
-  margin-bottom: 2%;
-  margin-top: 15%;
-  text-align: center;
-  width: 90px;
-  height: 90px;
+.boxButtonSide {
+  position: absolute;
+  top: 85%;
+  left: 90%;
+  width: 70px;
+  height: 70px;
   background-color: #19334d;
-  color: white;
-  font: Arial;
-  font-weight: bold;
-  font-size: 15px;
   border-style: none;
   border-radius: 50%;
 }
-img {
-  width: 60px;
+.boxButtonImg {
+  width: 50px;
 }
