+++ src/front-end/buttons/allContentButtons.js
@@ -16,8 +16,8 @@ class Buttons extends React.Component {
         <button className="split">Split CP</button>
         <button className="createFilter">Create Filter</button>
         <button className="editFilter">Edit Filter</button>
-        <button className="boxButton">
-          <img src="assets/box-icon.png" alt="Box-Button" />
+        <button className="boxButtonSide">
+          <img className="boxButtonImg" src="assets/box-icon.png" alt="Box-Button" />
         </button>
       </div>
     );
