+++ src/front-end/dropDown/dropDown.css
@@ -2,7 +2,7 @@
   font-family: sans-serif;
   background-color: #fff;
   margin-left: 86%;
-  width: 8%;
+  width: 10%;
   height: 2.8%;
   position: absolute;
   top: 2%;
@@ -62,9 +62,9 @@
 .dropdown-select.open {
   position: fixed;
   max-height: 50px;
-  width: 12%;
+  width: 10%;
   margin-left: 86.5%;
-  margin-top: 0.2%;
+  margin-top: 0.7%;
 }
 
 .dropdown-option {
