+++ src/front-end/header/header.css
@@ -1,6 +1,6 @@
 h1 {
   color: #0099ff;
-  padding-left: 13%;
+  padding-left: 14%;
   border-bottom: 4px solid #0099ff;
   font-family: sans-serif;
   padding-bottom: 5px;
