+++ src/front-end/input/allContentInput.css
@@ -1,5 +1,5 @@
 form {
-  padding-left: 13%;
+  padding-left: 15%;
   font-family: sans-serif;
 }
 input {
